% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_marker_names.R
\name{clean_marker_names}
\alias{clean_marker_names}
\title{Clean marker naming conventions}
\usage{
clean_marker_names(dat)
}
\arguments{
\item{dat}{dataframe: markers in columns and samples in rows}
}
\value{
dataframe with updated column names
}
\description{
Purpose: cleans column names to consistent formatting of
PERCENT_\if{html}{\out{<marker1>}}\emph{POS/NEG_OF}\if{html}{\out{<marker2>}}_POS/NEG
to match the naming in the count matrix created from the pipeline
}
\details{
Inputs:
}
