% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density_gates.R
\name{get_density_gates}
\alias{get_density_gates}
\title{Density gating of intensity values in \code{marker} for each unique subset of \code{subset_col}}
\usage{
get_density_gates(
  intens_dat,
  marker,
  subset_col,
  bin_n = 512,
  peak_detect_ratio = 10,
  pos_peak_threshold = 1600
)
}
\arguments{
\item{intens_dat}{dataframe of pre-gated (compensated, biexp. transf, gated CD4/CD8) intensity values where
cols = intensity value per marker,
rows = each sample}

\item{marker}{string for the marker to gate on
the name needs to match exactly the column name in \code{intens_dat}}

\item{subset_col}{string for the column name to indicate the subsets to apply density gating on
will perform operation on subsets corresponding to each unique value in column}

\item{bin_n}{numeric to be passed to \code{n} parameter of \code{density(n = bin_n)} for
number of equally spaced points at which the density is to be estimated
default is 512, which is the default of \code{density(n = 512)}}

\item{peak_detect_ratio}{numeric threshold for eliminating small peaks where
a peak that is < than the highest peak by \code{peak_detect_ratio} times will be ignored
default = 10}

\item{pos_peak_threshold}{numeric for threshold to identify a positive peak
'  default is 1600, which is on the biexponential scale}
}
\value{
tibble of gates/cutoffs for \code{marker} for each unique subset found in \code{subset_col}
rows correspond to unique values in \code{subset_col}, cols correspond to \code{marker}
}
\description{
For each unique value in \code{subset_col}, gate using density and estimated derivatives
to identify cutoff at shoulder (i.e., point of tapering off) after peak for
\code{marker} (intensity values)
}
