% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc.R
\name{get_perc}
\alias{get_perc}
\title{Calculate the percentage of positive cells for specific subpopulations}
\usage{
get_perc(
  intens_dat,
  num_marker,
  denom_marker,
  expand_num = FALSE,
  expand_denom = FALSE
)
}
\arguments{
\item{intens_dat}{dataframe of gated data with indicator columns per marker of interest
(specify in \code{num_marker} and \code{denom_marker})
with naming convention \code{marker_pos} per marker with
values of 0 to indicate negative-, 1 to indicate positive-expressing}

\item{num_marker}{string for the marker(s) to specify the numerator for subpopulations of interest
See \code{expand_num} argument and examples for how to specify}

\item{denom_marker}{string for the marker(s) to specify the denominator for subpopulations of interest
See \code{expand_denom} argument and examples for how to specify.}

\item{expand_num}{logical, only accepts \code{TRUE} or \code{FALSE} with default of \code{FALSE}
if \code{expand_num = TRUE}, currently only considers up to pairs of markers
specified in \code{num_marker} in the numerator of subpopulation calculations
(e.g., CD4+ & CD8- of CD3+)
if \code{expand_num = FALSE}, only considers each marker specified in \code{num_marker}
individually in the numerator of subpopulation calculations
(e.g., CD4+ of CD3+)}

\item{expand_denom}{logical, only accepts \code{TRUE} or \code{FALSE} with default of \code{FALSE}
if \code{expand_denom = TRUE}, currently considers up to 1 marker from the \code{num_marker} and
the unique combinations of \code{denom_marker} to generate list of subpopulations
(e.g., if \code{denom_marker = c("CD8")}, \code{num_marker = c("LAG3", "KI67")}, and \code{expand_denom = TRUE},
the subpopulations will include:
1. LAG3+ of CD8+, LAG3- of CD8+, LAG3+ of CD8-, LAG3- of CD8-,
2. KI67+ of CD8+, KI67- of CD8+, KI67+ of CD8-, KI67- of CD8-,
3. KI67+ of (LAG3+ & CD8+), KI67- of (LAG3+ & CD8+), KI67+ of (LAG3+ & CD8-), KI67- of (LAG3+ & CD8-)...etc.,
4. LAG3+ of (KI67+ & CD8+), LAG3- of (KI67+ & CD8+), LAG3+ of (KI67+ & CD8-), LAG3- of (KI67+ & CD8-)...etc.,
if \code{expand_denom = FALSE}, only generates the list of subpopulations based on
unique combinations of the \code{denom_marker}
(e.g., \code{denom_marker = c("CD4")} and \code{expand_denom = FALSE} only considers subpopulations with
denominator CD4+ and CD4- whereas
\verb{denom_marker = c("CD4", "CD8"} and \code{expand_denom = FALSE} will consider subpopulations with denominators
(CD4- & CD8-), (CD4+ & CD8-), (CD4- & CD8+) and (CD4+ & CD8+))}

\item{keep_indicators}{logical, only accepts \code{TRUE} or \code{FALSE} with default of \code{FALSE}
if \code{keep_indicators = TRUE}, will return indicator columns of 0/1 to specify which markers are considered in the
numerator and denominators of the subpopulations. This is useful for matching to percentage data
with potentially different naming conventions to avoid not having exact string matches for the same
subpopulation
take note that the order also matters when matching strings: "CD4+ & CD8- of CD3+" is different from "CD8- & CD4+ of CD3+"}
}
\value{
tibble containing the percentage of cells where
rows correspond to each subpopulation
}
\description{
Expects data input same as the output from \code{get_gated_dat} with indicator columns of specific
naming convention (see below).
}
\details{
The subpopulations are defined as (num marker(s)) out of (denom marker(s)) where
num denotes numerator, and denom denotes denominator
(these shorthands are used in the function arguments)
}
