% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompGS.R
\name{getCompGS}
\alias{getCompGS}
\title{Applies Compensation using specifications in csv file provided at \code{path_comp_mat}}
\usage{
getCompGS(gs, path_comp_mat)
}
\arguments{
\item{gs}{GatingSet to apply Biexponential Transformation to}

\item{path_comp_mat}{file path for .csv file that specifies the Compensation Matrix}
}
\value{
GatingSet with compensated data
}
\description{
The csv file at \code{path_comp_mat} should specify the channels to apply the compensation to.
The format is a matrix where the col and row names correspond to the channel names
}
\details{
An example is provided in the extdata/comp_mat_example_fcs.csv
}
