% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGatedDat.R
\name{getGatedDat}
\alias{getGatedDat}
\title{Attach indicator columns to \code{intens_dat} based on gates provided in \code{cutoffs}}
\usage{
getGatedDat(intens_dat = intensity_dat, cutoffs, subset_col)
}
\arguments{
\item{intens_dat}{dataframe of pre-gated (compensated, biexp. transf, openCyto steps) intensity values where
rows = each sample}

\item{cutoffs}{tibble of gates/cutoffs for all markers to gate
Expects \code{cutoffs} to match format of output from \code{\link[=getDensityGates]{getDensityGates()}} with
column corresponding to a marker, and rows to the subsets defined in the
\code{subset_col}}

\item{subset_col}{string for the column name to indicate the subsets to apply density gating on
will perform operation on subsets corresponding to each unique value in column}
}
\value{
\code{intens_dat} with additional cols attached for each marker in \code{cutoffs}
}
\description{
Adds an indicator column (0/1) to \code{intens_dat} for each marker in \code{cutoffs}
as indicated by the columns in \code{cutoffs}
}
\details{
The naming convention for the tagged on indicator columns will be
\verb{tolower(<marker_name>_pos)} where
0 indicates negativity or intensity < gate provided
1 indicates positivity or intensity > gate provided
}
