% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiexpTransformGS.R
\name{getBiexpTransformGS}
\alias{getBiexpTransformGS}
\title{Applies Biexpeonential Transformation using specifications in csv file provided at \code{path_biexp_params}}
\usage{
getBiexpTransformGS(gs, path_biexp_params)
}
\arguments{
\item{gs}{GatingSet to apply Biexponential Transformation to}

\item{path_biexp_params}{file path for .csv file that specifies the Biexponential Transformation}
}
\value{
GatingSet with Biexponentially Transformed data
}
\description{
The csv file at \code{path_biexp_params} should specify the channels to apply the transformation to
and the parameters (negative decades, width basis and positive decades).
}
\details{
The default is negative decades = 0.5, width basis = -30 and positive decades = 4.5.

The Transformation can be applied to only a subset of the channels included in the
GatingSet.

An example table is provided in the extdata/biexp_transf_parameters_x50.csv
}
\examples{
\dontrun{
path_biexp_params = system.file("extdata", "biexp_transf_parameters_x50.csv",
                                 package = "staRgate", mustWork = TRUE)

# gs must be a GatingSet object
gs = getBiexpTransformGS(gs, path_biexp_params = path_biexp_params)
}

}
