% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getDensityDerivs}
\alias{getDensityDerivs}
\title{Internal function: Estimate derivatives for density of \code{marker} for each unique subset of \code{subset_col}}
\usage{
getDensityDerivs(dens)
}
\arguments{
\item{dens}{\code{density} object from the \link[stats]{density}}
}
\value{
list of dataframe with density estimation and corresponding 1st-4th derivatives,
indicators of local peaks, plateau_pre \cr
each element corresponds to each unique value of \code{subset_col} \cr
for each dataframe: rows correspond to each of the bins
}
\description{
Internal function for \code{get_density_gates}
For each unique value in \code{subset_col}, estimate the derivatives for
\code{marker} (intensity values)
}
\keyword{internal}
