% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRgate-package.R
\docType{package}
\name{staRgate-package}
\alias{staRgate}
\alias{staRgate-package}
\title{staRgate: Automated gating pipeline for flow cytometry analysis to characterize the lineage, differentiation, and functional states of T-cells}
\description{
An R-based automated gating pipeline for flow cytometry data (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leejasme.github.io/staRgate/}
}

}
\author{
\strong{Maintainer}: Jasme Lee \email{leej22@mskcc.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
