% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRgate-package.R
\docType{package}
\name{staRgate-package}
\alias{staRgate}
\alias{staRgate-package}
\title{staRgate: Automated gating pipeline for flow cytometry analysis to characterize the lineage, differentiation, and functional states of T-cells}
\description{
An R-based automated gating pipeline for flow cytometry data designed to mimic the manual gating strategy of defining flow biomarker positive populations relative to a unimodal background population to include cells with varying intensities of marker expression. The pipeline’s main feature is a flexible density-based gating strategy capable of capturing varying scenarios based on marker expression patterns to analyze a 29-marker flow panel that characterizes T-cell lineage, differentiation, and functional states.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leejasme.github.io/staRgate/}
  \item \url{https://github.com/leejasme/staRgate}
  \item Report bugs at \url{https://github.com/leejasme/staRgate/issues}
}

}
\author{
\strong{Maintainer}: Jasme Lee \email{leej22@mskcc.org} (\href{https://orcid.org/0009-0006-4492-4872}{ORCID})

Authors:
\itemize{
  \item Matthew Adamow
  \item Colleen Maher
  \item Xiyu Peng
  \item Phillip Wong
  \item Fiona Ehrich
  \item Michael A Postow
  \item Margaret K Callahan
  \item Ronglai Shen
  \item Katherine S Panageas
}

}
\keyword{internal}
