% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRgate-package.R
\docType{package}
\name{staRgate-package}
\alias{staRgate}
\alias{staRgate-package}
\title{staRgate: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jasme Lee \email{leej22@mskcc.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
